[修正]拖曳元素至 pdf 位置錯誤問題
